# Further distance transform speedups.