Before:
```
running 2 tests
test benches::bench_chebyshev_distance_transform ... bench:       9,446 ns/iter (+/- 556)
test benches::bench_manhattan_distance_transform ... bench:       5,658 ns/iter (+/- 141)
```

After:
```
running 2 tests
test benches::bench_chebyshev_distance_transform ... bench:       7,340 ns/iter (+/- 267)
test benches::bench_manhattan_distance_transform ... bench:       4,401 ns/iter (+/- 143)
```

Replaces zipping the `RoomXY` iterators with a simple map for the column shift, and makes extensive use of checked arithmetic to avoid unnecessary writes.